requirements.txt
Pillow==7.2.0
psutil==5.7.2
requests-html==0.10.0
    bs4==0.0.1
        beautifulsoup4==4.9.1
            soupsieve==2.0.1
    fake-useragent==0.1.11
    parse==1.16.0
    pyppeteer==0.2.2
        appdirs==1.4.4
        pyee==7.0.2
        tqdm==4.48.2
        urllib3==1.25.10
        websockets==8.1
    pyquery==1.4.1
        cssselect==1.1.0
        lxml==4.5.2
    requests==2.24.0
        certifi==2020.6.20
        chardet==3.0.4
        idna==2.10
    w3lib==1.22.0
        six==1.15.0
setuptools==49.6.0
pytest==6.0.1
    atomicwrites==1.4.0
    attrs==20.1.0
    colorama==0.4.3
    iniconfig==1.0.1
    more-itertools==8.4.0
    packaging==20.4
        pyparsing==2.4.7
    pluggy==0.13.1
    py==1.9.0
    toml==0.10.1

